Add lot merge to unbook_postings.
`group_postings_by_meta` is used in description_based_source and schwab_csv
to return postings to the original form that the source generated so the
source can properly deduplicate postings.

This works well for automatic booking splits, such as those done by
FIFO, but doesn't work for manually split lots (for example high cost
first).

With this patch, one can now add `merge: TRUE` to a posting's metadata
and it will be merged with the posting one line above it.